requirements.txt
bert-for-tf2
focal-loss
tensorflow-addons
python-telegram-bot